(feat) profiles: Changed doubles to int

+++ lib/models/profiles_model.dart
@@ -1,6 +1,6 @@
 class ProfilesModel {
   // Saves items used
-  double userID = 0;
+  int userID = 0;
   List<bool> isAvailable = [false, false, false, false, false];
   // Constructor
   ProfilesModel({required this.userID, List<bool>? isAvailable})
@@ -9,14 +9,14 @@ class ProfilesModel {
   List<bool> getEquippedItems(){
     return isAvailable;
   }
-  double getUserID(){
+  int getUserID(){
     return userID;
   }
   //setters
   void setItemAvailable(int itemID, bool value){
     isAvailable[itemID] = value;
   }
-  void setUserID(double newID){
+  void setUserID(int newID){
     userID = newID;
   }
 }
